refactor(migrations): change function return type from decimal to bigint

+++ migrations/20250402012629_add_account_view.sql
@@ -7,7 +7,7 @@ returns table (
     category text,
     description text,
     type text,
-    amount decimal
+    amount bigint
 ) as $$
 begin
     -- return transactions for the specified account
